+++ CHANGELOG.md
@@ -7,12 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
+### Fixed
+- Fix repository connection ensure (if callback throw)
 
 ## [0.4.4] - 2024-07-22
-
 ### Fixed
-
 - nikic/php-parser 5 support
 
 ## [0.4.3] - 2024-04-25