+++ CHANGELOG.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- nette/database 3.2 support
+- PHP8.3 required
+- sorting to top/bottom
 ### Fixed
 - Fix repository connection ensure (if callback throw)
-## Added
-- sorting to top/bottom
 
 ## [0.4.4] - 2024-07-22
 ### Fixed